feat(codetable-api): 抽出 AbstractCodeTableMutationHandler 基底（Phase B 起手）
Code 表受審計寫入 API 的第一步（CODE_TABLE_MUTATION_API_PLAN §4/§5 步驟 3-4）：把
NianHaoMutationHandler 的交易・audit_log・operations・複合主鍵驗證・欄位白名單・變更偵測・
direct/proposal 樣板邏輯抽到共用基底 AbstractCodeTableMutationHandler，具體每表 handler 只需
宣告 tableName／resourceName／resourceAliases／displayName／keyColumns／allowedFields。

- 行為對 NIAN_HAO 完全等價（22 個 ApiV2MutateNianHao 測試不變）。
- 泛化支援複合主鍵（whereByPk 對全鍵 AND，供日後 TEXT_INSTANCE_DATA 等 3 鍵表）。
- 防呆：keyColumns() 必須與 CompositePrimaryKey::SCHEMAS 登錄一致，否則 500；杜絕子類宣告部分鍵
  導致 UPDATE 命中多列的風險（review agent 揪出的 future footgun，已硬擋）。
- code 表 c_personid 一律 0（沿用 NianHao 既有做法，MutationController 零改動、§D-3 選項 b）。

review agent + codex 各一輪，均無 SERIOUS。
